refactor: move a2p settlement contract to enterprise build
- update build workflow
- fix RPC test case
- update code owner
- fix #1146
