+++ .github/workflows/go.yml
@@ -8,7 +8,7 @@ jobs:
     if: contains(github.event.head_commit.message, '[skip ci]') == false && contains(github.event.head_commit.message, '[ci skip]') == false
     steps:
       - name: Set up Go
-        uses: actions/setup-go@v1.1.2
+        uses: actions/setup-go@v2
         with:
           go-version: 1.14.x
 
@@ -17,7 +17,7 @@ jobs:
         with:
           fetch-depth: 1
 
-      - uses: actions/cache@v1.1.1
+      - uses: actions/cache@v2
         with:
           path: ~/go/pkg/mod
           key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
@@ -33,11 +33,11 @@ jobs:
             gofmt -d .
             exit 1
           fi
-          go vet `go list ./... | egrep -v "platform"`
+          go vet `go list ./...`
           go vet -tags testnet -v `go list ./...`
           echo "" > "coverage.txt"
           for d in $(go list ./... | egrep -v 'mocks'); do
-            go test -coverprofile=profile.out -covermode=atomic "$d"
+            go test -tags testnet -coverprofile=profile.out -covermode=atomic "$d"
             if [[ -f profile.out ]]; then
               cat profile.out >> "coverage.txt"
               rm profile.out
