+++ rpc/grpc/apis/settlement.go
@@ -1,3 +1,5 @@
+// +build testnet
+
 package apis
 
 import (
