+++ vm/contract/settlement.go
@@ -1,3 +1,5 @@
+// +build testnet
+
 /*
  * Copyright (c) 2019 QLC Chain Team
  *
